(CLOUD-403) initial skeleton for azure_vm provider
This implements the smallest number of properties to allow for creating
and destroying a vm. It doesn't attempt to expand this to the full type,
or add any additional read-only properties.

Note that this swaps in the prerelease version of the Azure SDK. This
purports to offer better Windows support as well as has a new module
interface so where starting with that seems prudent.