(CLOUD-404) Support credentials in files as well as ENV
Based on user feedback, and also on the act the Puppet agent tends to
clober ENV variables, it's useful to support configuration in files.
It's also useful to support ENV vars for both development and for more
interactive usage.

This commit adds a simple tested interface for accessing configuration,
and supports the absolute bare minimum properties to access the Azure
Virtual Machine API. It's likely we'll add more properties in the future
but this should provide a simple way of doing so.