(CLOUD-425) Support location property
Location is useful as a property so we can read it back in puppet
resource, but VMs can't be moved simply between locations - so we have
made location read-only after creation.

Validation of the region is left to the API, so as to avoid encoding a
list of locations in the module which will change over time.