(CLOUD-417) Acceptance test full master/agent support
The acceptance tests for azure now support a full master/agent
setup, as well as agent only and local testing setup. The master
agent performs the tests after doing a puppet agent run on the agent.
The agent only runs a puppet apply. The ENV variable
PUPPET_AZURE_BEAKER_MODE is responsible for switching the tests. Valid
settings are 'agent', 'apply' or 'local'. If this variable is not set,
it will default to 'apply'.