(MAINT) acceptance helper: deploy the module even if BEAKER_provision=no
install_pe and friends set the SUT machines up and only need to be run
once for a nodeset. Re-using the nodes with BEAKER_destroy=no and
BEAKER_provision=no doesn't re-run this.

On the other hand, the puppet_module_install deploys a snapshot of the
local code to the SUT and needs to be run every time to ensure that we
are actually testing our latest work.