(MAINT) Remove the forge stub to avoid downloading wrong module
install_dev_puppet_module_on will use the internal forge if this
variable is set. This isn't actually the behaviour we want locally,
although it probably is in the CI pipeline eventually.

Also, due to the fact their is already a puppetlabs-azure module that
needs replacing we run into entertaining naming issues. For the moment
removing it feels the best approach.