(CLOUD-414) Document how to get the PEM file for use with Puppet
I found getting hold of the credentials non-obvious, in that you don't
appear to be able to do any of this from within the Azure console. The
official documentation is actually pretty good, but it's very specific
so you need to read multiple guides. This change documents the
requirements from the point of view of the happy path for Puppet users.