(CLOUD-438) Support the resource management API in Azure
Azure currently has two APIs; service management (or classic) and
resource management (or ARM for short). This commit adds initial support
for the resource management API in the form of the azure_vm type.

Specifically we support a pretty low level interface to a wide range of
properties for the VM, as well as the ability to create, stop, restart
and destroy machines. The type does not yet support any form of
mutation.