(CLOUD-476) UI improvements to the azure_vm type
This PR makes a number of improvements to the user interface for the
azure_vm type. In particular:

* Adds some size validations. Ideally this would come from the API, but
the exceptions from the ARM API are not user friendly. Long names are
the easies way of triggered a user-unfriendly error.
* Sets the defaults for the type to match those of the Azure Portal, this
should better align with user expectations
* Mirrored some of the API behaviour around case sensitivity in the type,
to avoid out of sync issues
* Provides a minimal example, along with a matching acceptance test
Updated relevant tests and documentation