(Docs) Updating Documentation for New Azure Types (#89)
* (FM-4516) (Docs) Add extensions to azure_vm

Adding documentation to the README to define behavior and parameters on
the Azure VM extension. Also, provided a working example (tested 6/20)
in the README to show how to use the extension.

* (FM-5230) (Docs) Documentation for Storage Accounts

Adding documentation for new Storage Account type and provider. This
provides a detailed list of parameters as well as working example of
how to provision a Storage Account. Also, made formatting across
parameters consistent to ensure working preview across most markdown
editors.

* (FM-5229)(Docs) Documentation for Resource Groups

Adding documentation and working example for Azure Resource Group.
Identifies allowed parameters and links to Azure documentation.

* (Docs) Adding More Description to Extension Docs

Adding more description to extension docs to identify an example in the
reference section as well as link to the Azure documentation.

* (Docs) Combining Extension Docs Improvements

Adding all the extension docs improvements into a single commit.