Fix test of customdata on azure_vm
Previously the test tried to use `@machine.ip` to access the public IP
address of the newly created azure instance, just like the similar test
in `azure_vm_classic`. This doesn't work, so we need new methods to find
the public ip address of the azure ARM instance before the test can pass.