(MODULES-5005) Support network security group name
Add a new parameter `network_security_group_name` that allows the network security group name to be set at time of VM creation.  If security group already exists, it will be attached to the VM, otherwise it will be created.  If the special name `None` is passed, then no security group will be created or associated.  Due to the way azure API orders resources, its not optimal to inspect the VM and return the security group name that applies (security groups belong to network interfaces not VMs), so the `puppet resource` command does not interrogate the current resource status.  This means that the Security Group cannot currently be changed by puppet once set.  This decision is inline with other aspects of `azure_vm` that are only used at creation time.